+++ changelog.md
@@ -2,4 +2,5 @@
 
 ## Version 0.1.0 (2023/02/19)
 
+- Added `list2json` command
 - Init